+++ chromium/media/video/gpu_memory_buffer_video_frame_pool.cc
@@ -16,6 +16,7 @@
 
 #include "base/barrier_closure.h"
 #include "base/bind.h"
+#include "base/containers/circular_deque.h"
 #include "base/containers/stack_container.h"
 #include "base/location.h"
 #include "base/macros.h"
@@ -71,12 +72,15 @@ class GpuMemoryBufferVideoFramePool::PoolImpl
   bool OnMemoryDump(const base::trace_event::MemoryDumpArgs& args,
                     base::trace_event::ProcessMemoryDump* pmd) override;
 
+  // Aborts any pending copies.
+  void Abort();
+
   // Shuts down the frame pool and releases all frames in |frames_|.
   // Once this is called frames will no longer be inserted back into
   // |frames_|.
   void Shutdown();
 
-  void SetTickClockForTesting(base::TickClock* tick_clock);
+  void SetTickClockForTesting(const base::TickClock* tick_clock);
 
  private:
   friend class base::RefCountedThreadSafe<
@@ -117,25 +121,40 @@ class GpuMemoryBufferVideoFramePool::PoolImpl
     base::TimeTicks last_use_time_;
   };
 
+  // Struct to keep track of requested videoframe copies.
+  struct VideoFrameCopyRequest {
+    VideoFrameCopyRequest(scoped_refptr<VideoFrame> video_frame,
+                          FrameReadyCB frame_ready_cb,
+                          bool passthrough)
+        : video_frame(video_frame),
+          frame_ready_cb(std::move(frame_ready_cb)),
+          passthrough(passthrough) {}
+    scoped_refptr<VideoFrame> video_frame;
+    FrameReadyCB frame_ready_cb;
+    bool passthrough;
+  };
+
+  // Start the copy of a video_frame on the worker_task_runner_.
+  // It assumes there are currently no in-flight copies and works on the request
+  // in the front of |frame_copy_requests_| queue.
+  void StartCopy();
+
   // Copy |video_frame| data into |frame_resources| and calls |frame_ready_cb|
   // when done.
   void CopyVideoFrameToGpuMemoryBuffers(
       const scoped_refptr<VideoFrame>& video_frame,
-      FrameResources* frame_resources,
-      FrameReadyCB frame_ready_cb);
+      FrameResources* frame_resources);
 
   // Called when all the data has been copied.
   void OnCopiesDone(const scoped_refptr<VideoFrame>& video_frame,
-                    FrameResources* frame_resources,
-                    FrameReadyCB frame_ready_cb);
+                    FrameResources* frame_resources);
 
   // Prepares GL resources, mailboxes and calls |frame_ready_cb| with the new
   // VideoFrame. This has to be run on |media_task_runner_| where
-  // |frame_ready_cb| will also be run.
+  // |frame_ready_cb| associated with video_frame will also be run.
   void BindAndCreateMailboxesHardwareFrameResources(
       const scoped_refptr<VideoFrame>& video_frame,
-      FrameResources* frame_resources,
-      FrameReadyCB frame_ready_cb);
+      FrameResources* frame_resources);
 
   // Return true if |resources| can be used to represent a frame for
   // specific |format| and |size|.
@@ -150,6 +169,13 @@ class GpuMemoryBufferVideoFramePool::PoolImpl
   FrameResources* GetOrCreateFrameResources(
       const gfx::Size& size,
       GpuVideoAcceleratorFactories::OutputFormat format);
+
+  // Calls the FrameReadyCB of the first entry in |frame_copy_requests_|, with
+  // the provided |video_frame|, then deletes the entry from
+  // |frame_copy_requests_| and attempts to start another copy if there are
+  // other |frame_copy_requests_| elements.
+  void CompleteCopyRequestAndMaybeStartNextCopy(
+      const scoped_refptr<VideoFrame>& video_frame);
 
   // Callback called when a VideoFrame generated with GetFrameResources is no
   // longer referenced.
@@ -177,8 +203,11 @@ class GpuMemoryBufferVideoFramePool::PoolImpl
   GpuVideoAcceleratorFactories::OutputFormat output_format_;
 
   // |tick_clock_| is always a DefaultTickClock outside of testing.
-  base::TickClock* tick_clock_;
+  const base::TickClock* tick_clock_;
 
+  // Queued up video frames for copies. The front is the currently
+  // in-flight copy, new copies are added at the end.
+  base::circular_deque<VideoFrameCopyRequest> frame_copy_requests_;
   bool in_shutdown_;
 
   DISALLOW_COPY_AND_ASSIGN(PoolImpl);
@@ -525,10 +554,9 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::CreateHardwareFrame(
         gpu_factories_->VideoFrameOutputFormat(video_frame->BitDepth());
   }
 
-  if (output_format_ == GpuVideoAcceleratorFactories::OutputFormat::UNDEFINED) {
-    std::move(frame_ready_cb).Run(video_frame);
-    return;
-  }
+  bool passthrough = false;
+  if (output_format_ == GpuVideoAcceleratorFactories::OutputFormat::UNDEFINED)
+    passthrough = true;
   switch (video_frame->format()) {
     // Supported cases.
     case PIXEL_FORMAT_YV12:
@@ -559,23 +587,23 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::CreateHardwareFrame(
     case PIXEL_FORMAT_YUV444P12:
     case PIXEL_FORMAT_Y16:
     case PIXEL_FORMAT_UNKNOWN:
-      std::move(frame_ready_cb).Run(video_frame);
-      return;
+      passthrough = true;
+  }
+  // TODO(dcastagna): Handle odd positioned video frame input, see
+  // https://crbug.com/638906.
+  // TODO(emircan): Eliminate odd size video frame input cases as they are not
+  // valid, see https://crbug.com/webrtc/9033.
+  if ((video_frame->visible_rect().x() & 1) ||
+      (video_frame->visible_rect().y() & 1) ||
+      (video_frame->coded_size().width() & 1) ||
+      (video_frame->coded_size().height() & 1)) {
+    passthrough = true;
   }
 
-  const gfx::Size coded_size = CodedSize(video_frame, output_format_);
-  // Acquire resources. Incompatible ones will be dropped from the pool.
-  FrameResources* frame_resources =
-      GetOrCreateFrameResources(coded_size, output_format_);
-  if (!frame_resources) {
-    std::move(frame_ready_cb).Run(video_frame);
-    return;
-  }
-
-  worker_task_runner_->PostTask(
-      FROM_HERE, base::BindOnce(&PoolImpl::CopyVideoFrameToGpuMemoryBuffers,
-                                this, video_frame, frame_resources,
-                                base::Passed(&frame_ready_cb)));
+  frame_copy_requests_.emplace_back(video_frame, std::move(frame_ready_cb),
+                                    passthrough);
+  if (frame_copy_requests_.size() == 1u)
+    StartCopy();
 }
 
 bool GpuMemoryBufferVideoFramePool::PoolImpl::OnMemoryDump(
@@ -621,10 +649,19 @@ bool GpuMemoryBufferVideoFramePool::PoolImpl::OnMemoryDump(
   return true;
 }
 
+void GpuMemoryBufferVideoFramePool::PoolImpl::Abort() {
+  DCHECK(media_task_runner_->BelongsToCurrentThread());
+  // Abort any pending copy requests. If one is already in flight, we can't do
+  // anything about it.
+  if (frame_copy_requests_.size() <= 1u)
+    return;
+  frame_copy_requests_.erase(frame_copy_requests_.begin() + 1,
+                             frame_copy_requests_.end());
+}
+
 void GpuMemoryBufferVideoFramePool::PoolImpl::OnCopiesDone(
     const scoped_refptr<VideoFrame>& video_frame,
-    FrameResources* frame_resources,
-    FrameReadyCB frame_ready_cb) {
+    FrameResources* frame_resources) {
   for (const auto& plane_resource : frame_resources->plane_resources) {
     if (plane_resource.gpu_memory_buffer) {
       plane_resource.gpu_memory_buffer->Unmap();
@@ -639,8 +676,33 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::OnCopiesDone(
   media_task_runner_->PostTask(
       FROM_HERE,
       base::BindOnce(&PoolImpl::BindAndCreateMailboxesHardwareFrameResources,
-                     this, video_frame, frame_resources,
-                     base::Passed(&frame_ready_cb)));
+                     this, video_frame, frame_resources));
+}
+
+void GpuMemoryBufferVideoFramePool::PoolImpl::StartCopy() {
+  DCHECK(media_task_runner_->BelongsToCurrentThread());
+  DCHECK(!frame_copy_requests_.empty());
+
+  while (!frame_copy_requests_.empty()) {
+    VideoFrameCopyRequest& request = frame_copy_requests_.front();
+    // Acquire resources. Incompatible ones will be dropped from the pool.
+    FrameResources* frame_resources =
+        request.passthrough
+            ? nullptr
+            : GetOrCreateFrameResources(
+                  CodedSize(request.video_frame, output_format_),
+                  output_format_);
+    if (!frame_resources) {
+      std::move(request.frame_ready_cb).Run(request.video_frame);
+      frame_copy_requests_.pop_front();
+      continue;
+    }
+
+    worker_task_runner_->PostTask(
+        FROM_HERE, base::BindOnce(&PoolImpl::CopyVideoFrameToGpuMemoryBuffers,
+                                  this, request.video_frame, frame_resources));
+    break;
+  }
 }
 
 // Copies |video_frame| into |frame_resources| asynchronously, posting n tasks
@@ -648,8 +710,7 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::OnCopiesDone(
 // After the barrier is passed OnCopiesDone will be called.
 void GpuMemoryBufferVideoFramePool::PoolImpl::CopyVideoFrameToGpuMemoryBuffers(
     const scoped_refptr<VideoFrame>& video_frame,
-    FrameResources* frame_resources,
-    FrameReadyCB frame_ready_cb) {
+    FrameResources* frame_resources) {
   // Compute the number of tasks to post and create the barrier.
   const size_t num_planes = VideoFrame::NumPlanes(VideoFormat(output_format_));
   const size_t planes_per_copy = PlanesPerCopy(output_format_);
@@ -665,11 +726,9 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::CopyVideoFrameToGpuMemoryBuffers(
       ++copies;
   }
 
-  const base::Closure copies_done =
-      base::Bind(&PoolImpl::OnCopiesDone, this, video_frame, frame_resources,
-                 base::Passed(&frame_ready_cb));
-  const base::RepeatingClosure barrier =
-      base::BarrierClosure(copies, copies_done);
+  const base::RepeatingClosure barrier = base::BarrierClosure(
+      copies, base::BindOnce(&PoolImpl::OnCopiesDone, this, video_frame,
+                             frame_resources));
 
   // Map the buffers.
   for (size_t i = 0; i < NumGpuMemoryBuffers(output_format_); i++) {
@@ -763,16 +822,13 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::CopyVideoFrameToGpuMemoryBuffers(
 void GpuMemoryBufferVideoFramePool::PoolImpl::
     BindAndCreateMailboxesHardwareFrameResources(
         const scoped_refptr<VideoFrame>& video_frame,
-        FrameResources* frame_resources,
-        FrameReadyCB frame_ready_cb) {
-  std::unique_ptr<GpuVideoAcceleratorFactories::ScopedGLContextLock> lock(
-      gpu_factories_->GetGLContextLock());
-  if (!lock) {
+        FrameResources* frame_resources) {
+  gpu::gles2::GLES2Interface* gles2 = gpu_factories_->ContextGL();
+  if (!gles2) {
     frame_resources->MarkUnused(tick_clock_->NowTicks());
-    std::move(frame_ready_cb).Run(video_frame);
+    CompleteCopyRequestAndMaybeStartNextCopy(video_frame);
     return;
   }
-  gpu::gles2::GLES2Interface* gles2 = lock->ContextGL();
 
   const gfx::Size coded_size = CodedSize(video_frame, output_format_);
   gpu::MailboxHolder mailbox_holders[VideoFrame::kMaxPlanes];
@@ -809,24 +865,23 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::
   for (size_t i = 0; i < NumGpuMemoryBuffers(output_format_); i++)
     mailbox_holders[i].sync_token = sync_token;
 
-  auto release_mailbox_callback = BindToCurrentLoop(
-      base::Bind(&PoolImpl::MailboxHoldersReleased, this, frame_resources));
-
   VideoPixelFormat frame_format = VideoFormat(output_format_);
 
   // Create the VideoFrame backed by native textures.
   gfx::Size visible_size = video_frame->visible_rect().size();
   scoped_refptr<VideoFrame> frame = VideoFrame::WrapNativeTextures(
-      frame_format, mailbox_holders, release_mailbox_callback, coded_size,
+      frame_format, mailbox_holders, VideoFrame::ReleaseMailboxCB(), coded_size,
       gfx::Rect(visible_size), video_frame->natural_size(),
       video_frame->timestamp());
 
   if (!frame) {
     frame_resources->MarkUnused(tick_clock_->NowTicks());
-    release_mailbox_callback.Run(gpu::SyncToken());
-    std::move(frame_ready_cb).Run(video_frame);
+    MailboxHoldersReleased(frame_resources, gpu::SyncToken());
+    CompleteCopyRequestAndMaybeStartNextCopy(video_frame);
     return;
   }
+  frame->SetReleaseMailboxCB(
+      base::BindOnce(&PoolImpl::MailboxHoldersReleased, this, frame_resources));
 
   frame->set_color_space(video_frame->ColorSpace());
 
@@ -851,20 +906,14 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::
     default:
       break;
   }
+
+  frame->metadata()->MergeMetadataFrom(video_frame->metadata());
   frame->metadata()->SetBoolean(VideoFrameMetadata::ALLOW_OVERLAY,
                                 allow_overlay);
-
-  base::TimeTicks render_time;
-  if (video_frame->metadata()->GetTimeTicks(VideoFrameMetadata::REFERENCE_TIME,
-                                            &render_time)) {
-    frame->metadata()->SetTimeTicks(VideoFrameMetadata::REFERENCE_TIME,
-                                    render_time);
-  }
-
   frame->metadata()->SetBoolean(VideoFrameMetadata::READ_LOCK_FENCES_ENABLED,
                                 true);
 
-  std::move(frame_ready_cb).Run(frame);
+  CompleteCopyRequestAndMaybeStartNextCopy(frame);
 }
 
 // Destroy all the resources posting one task per FrameResources
@@ -874,6 +923,10 @@ GpuMemoryBufferVideoFramePool::PoolImpl::~PoolImpl() {
 }
 
 void GpuMemoryBufferVideoFramePool::PoolImpl::Shutdown() {
+  DCHECK(media_task_runner_->BelongsToCurrentThread());
+  // Clients don't care about copies once shutdown has started, so abort them.
+  Abort();
+
   // Delete all the resources on the media thread.
   in_shutdown_ = true;
   for (auto* frame_resources : resources_pool_) {
@@ -889,7 +942,7 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::Shutdown() {
 }
 
 void GpuMemoryBufferVideoFramePool::PoolImpl::SetTickClockForTesting(
-    base::TickClock* tick_clock) {
+    const base::TickClock* tick_clock) {
   tick_clock_ = tick_clock;
 }
 
@@ -899,7 +952,7 @@ GpuMemoryBufferVideoFramePool::PoolImpl::FrameResources*
 GpuMemoryBufferVideoFramePool::PoolImpl::GetOrCreateFrameResources(
     const gfx::Size& size,
     GpuVideoAcceleratorFactories::OutputFormat format) {
-  DCHECK(!in_shutdown_);
+  DCHECK(media_task_runner_->BelongsToCurrentThread());
 
   auto it = resources_pool_.begin();
   while (it != resources_pool_.end()) {
@@ -919,12 +972,10 @@ GpuMemoryBufferVideoFramePool::PoolImpl::GetOrCreateFrameResources(
   }
 
   // Create the resources.
-  std::unique_ptr<GpuVideoAcceleratorFactories::ScopedGLContextLock> lock(
-      gpu_factories_->GetGLContextLock());
-  if (!lock)
+  gpu::gles2::GLES2Interface* gles2 = gpu_factories_->ContextGL();
+  if (!gles2)
     return nullptr;
 
-  gpu::gles2::GLES2Interface* gles2 = lock->ContextGL();
   gles2->ActiveTexture(GL_TEXTURE0);
   FrameResources* frame_resources = new FrameResources(size);
   resources_pool_.push_back(frame_resources);
@@ -955,6 +1006,17 @@ GpuMemoryBufferVideoFramePool::PoolImpl::GetOrCreateFrameResources(
   return frame_resources;
 }
 
+void GpuMemoryBufferVideoFramePool::PoolImpl::
+    CompleteCopyRequestAndMaybeStartNextCopy(
+        const scoped_refptr<VideoFrame>& video_frame) {
+  DCHECK(!frame_copy_requests_.empty());
+
+  std::move(frame_copy_requests_.front().frame_ready_cb).Run(video_frame);
+  frame_copy_requests_.pop_front();
+  if (!frame_copy_requests_.empty())
+    StartCopy();
+}
+
 // static
 void GpuMemoryBufferVideoFramePool::PoolImpl::DeleteFrameResources(
     GpuVideoAcceleratorFactories* const gpu_factories,
@@ -962,12 +1024,9 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::DeleteFrameResources(
   // TODO(dcastagna): As soon as the context lost is dealt with in media,
   // make sure that we won't execute this callback (use a weak pointer to
   // the old context).
-
-  std::unique_ptr<GpuVideoAcceleratorFactories::ScopedGLContextLock> lock(
-      gpu_factories->GetGLContextLock());
-  if (!lock)
+  gpu::gles2::GLES2Interface* gles2 = gpu_factories->ContextGL();
+  if (!gles2)
     return;
-  gpu::gles2::GLES2Interface* gles2 = lock->ContextGL();
 
   for (PlaneResource& plane_resource : frame_resources->plane_resources) {
     if (plane_resource.image_id)
@@ -982,7 +1041,13 @@ void GpuMemoryBufferVideoFramePool::PoolImpl::DeleteFrameResources(
 void GpuMemoryBufferVideoFramePool::PoolImpl::MailboxHoldersReleased(
     FrameResources* frame_resources,
     const gpu::SyncToken& release_sync_token) {
-  DCHECK(media_task_runner_->BelongsToCurrentThread());
+  if (!media_task_runner_->BelongsToCurrentThread()) {
+    media_task_runner_->PostTask(
+        FROM_HERE, base::BindOnce(&PoolImpl::MailboxHoldersReleased, this,
+                                  frame_resources, release_sync_token));
+    return;
+  }
+
   if (in_shutdown_) {
     DeleteFrameResources(gpu_factories_, frame_resources);
     delete frame_resources;
@@ -1037,8 +1102,12 @@ void GpuMemoryBufferVideoFramePool::MaybeCreateHardwareFrame(
   pool_impl_->CreateHardwareFrame(video_frame, std::move(frame_ready_cb));
 }
 
+void GpuMemoryBufferVideoFramePool::Abort() {
+  pool_impl_->Abort();
+}
+
 void GpuMemoryBufferVideoFramePool::SetTickClockForTesting(
-    base::TickClock* tick_clock) {
+    const base::TickClock* tick_clock) {
   pool_impl_->SetTickClockForTesting(tick_clock);
 }
